+++ src/main/java/UnsisSmile/odonto/edu/UnsisSmile/model/IAdministradoresRepository.java
@@ -1,7 +1,6 @@
 package UnsisSmile.odonto.edu.UnsisSmile.model;
 
 import java.time.LocalDate;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.stereotype.Repository;
@@ -19,7 +18,7 @@ public interface IAdministradoresRepository  extends JpaRepository<Administrador
 	Administradores eliminarAdministrador(Long id_administrador);
 	
 	@Procedure(name= "obtenerAdministradores")
-	Administradores obtenAdministradores(int id_admin);
+	Administradores obtenAdministradores(Long id_admin);
 	
 	@Procedure(name="obtenerTodosAdministradores")
 	Administradores obtenerTodosAdministradores();
